+++ ThatGameCompanyGameJam/Assets/Scripts/ScriptableObjects/IngredientScriptableObject.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "DefaultIngredient", menuName = "IngredientSO", order = 0)]
 public class IngredientScriptableObject : ScriptableObject
 {
-    public float Cost;
+    public int Cost;
     // public Sprite Sprite;
     // Add type?
 }
